+++ TeamZenmaiSiki/Assets/Resources/Prefabs/Battle/EnemyModel.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 920a84e7062e3934d90778093b60454e
+timeCreated: 1477112082
+licenseType: Free
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
