--- TeamZenmaiSiki/Assets/Resources/Sprits/Search.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 5648cd96ffdeb904f9dc6bfbcf0a76b9
-folderAsset: yes
-timeCreated: 1477060257
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
