+++ TeamZenmaiSiki/Assets/Scripts/Mitui/Main/Battle/EnemyRead.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 1725cd63f839d04488d83af076a2fd79
-timeCreated: 1476669252
+guid: 52ba2ec0fc2066c41a56ada879b1dddf
+timeCreated: 1477277728
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
