--- TeamZenmaiSiki/Assets/Scripts/Mitui/Main/Battle/PlayerStatus.cs
@@ -1,14 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public class PlayerStatus : MonoBehaviour
-{
-    void OnTriggerEnter2D(Collider2D collision)
-    {
-        if (collision.tag == "AttackCircle")
-        {
-            Debug.Log("Hit!");
-            Destroy(this.gameObject);
-        }
-    }
-}
