--- TeamZenmaiSiki/Assets/Scripts/Mitui/Main/Battle/PlayerStatus.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: d95d536ad541bcb489348fea513b1286
-timeCreated: 1476669236
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
